Move fns to service from invoices controller
